+++ app/Modules/Dashboard/routes.php
@@ -13,4 +13,9 @@ Route::group(['middleware' => ['web', 'auth.admin'], 'namespace' => 'FI\Modules\
 {
     Route::get('/', 'DashboardController@index');
     Route::get('dashboard', ['uses' => 'DashboardController@index', 'as' => 'dashboard.index']);
+    //clear views from storage
+    Route::get('/viewclear', [function () {
+        Artisan::call('view:clear');
+        return redirect()->route('dashboard.index');
+    }]);
 });