+++ app/Modules/Payments/Views/payments/index.blade.php
@@ -39,6 +39,10 @@
 
             <a href="javascript:void(0)" class="btn btn-default bulk-actions" id="btn-bulk-delete"><i class="fa fa-trash"></i> {{ trans('fi.trash') }}</a>
 
+            <a href="javascript:void(0)" id="btn-enter-multi-payment" class="btn btn-primary enter-multi-payment"
+               data-redirect-to="{{ request()->fullUrl() }}"><i
+                        class="fa fa-credit-card"></i> {{ trans('fi.enter_payment') }}</a>
+
         </div>
 
         <div class="clearfix"></div>