+++ app/Modules/Quotes/Controllers/QuoteController.php
@@ -54,7 +54,7 @@ class QuoteController extends Controller
     {
         Quote::whereIn('id', request('ids'))->update(['quote_status_id' => request('status')]);
 
-        return response()->json(['success' => trans('status_successfully_updated')], 200);
+        return response()->json(['success' => trans('fi.status_successfully_updated')], 200);
     }
 
     public function pdf($id)