+++ app/Modules/Setup/Views/setup/migration.blade.php
@@ -47,9 +47,11 @@
                         <a class="btn btn-primary" id="btn-run-migration">{{ trans('fi.continue') }}</a>
 
                         <a class="btn btn-default" id="btn-running-migration" style="display: none;" disabled="disabled">{{ trans('fi.installing_please_wait') }}</a>
-
+                        @if($mode == 'setup')
                         <a href="{{ route('setup.neworxfer') }}" class="btn btn-success" id="btn-migration-complete" style="display: none;">{{ trans('fi.continue') }}</a>
-
+                        @else
+                        <a href="{{ route('setup.newaccount') }}" class="btn btn-success" id="btn-migration-complete" style="display: none;">{{ trans('fi.continue') }}</a>
+                        @endif
                     </div>
 
                 </div>