+++ database/migrations/2018_09_07_000100_version_401.php
@@ -1,0 +1,33 @@
+<?php
+
+use FI\Modules\Settings\Models\Setting;
+use Illuminate\Database\Migrations\Migration;
+
+class Version401 extends Migration
+{
+
+    /**
+     * Run the migrations.
+     * @table payments_custom
+     *
+     * @return void
+     */
+    public function up()
+    {
+        deleteTempFiles();
+        deleteViewCache();
+
+        Setting::saveByKey('version', '4.0.1');
+
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+     public function down()
+     {
+       //
+     }
+}