+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 4.0.1
+
+- added "todays workorder" widget
+- added "recent payments" widget
+- modified setup for upgrade
+- added resources/lang/en-cust where client = customer
+- added system info (.env settings) tab to system settings
+- added modal Enter Payment function, with client lookup and payable invoices
+
+
 ## 4.0.0
 
 - change server to public directory (requires apache change)